Reorganized the current tests as per C# standards
